remove old registry and PersistentModule
